# rest/config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Liverpool FC API",
    "version": "1.0.0",
    "description": "REST API for managing Liverpool Football Club data including players, matches, and more"
  },
  "servers": [
    {
      "url": "http://localhost:8989/api/v1"
    }
  ],
  "paths": {
...
    "/players": {
      "get": {
        "summary": "Get all players",
        "tags": [
          "Players"
        ],
        "responses": {
          "200": {
            "description": "List of players",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of players",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Player"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No players found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new player",
        "tags": [
          "Players"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Player created",
            "headers": {
              "Location": {
                "description": "URL of created player",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Jersey number already taken"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/players/{id}": {
      "get": {
        "summary": "Get player by ID",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Player found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update player",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Player not found"
          },
          "409": {
            "description": "Jersey number already taken"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Partially update player",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Player not found"
          },
          "409": {
            "description": "Jersey number already taken"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete player",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Player deleted"
          },
          "404": {
            "description": "Player not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/player-stats": {
      "get": {
        "summary": "Get all player stats",
        "tags": [
          "PlayerStats"
        ],
        "responses": {
          "200": {
            "description": "List of player stats",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of player stats records",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PlayerStats"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No player stats found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new player stats",
        "tags": [
          "PlayerStats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerStats"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Player stats created",
            "headers": {
              "Location": {
                "description": "URL of created player stats",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/player-stats/{id}": {
      "get": {
        "summary": "Get player stats by ID",
        "tags": [
          "PlayerStats"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "PlayerStats ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Player stats found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerStatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player stats not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update player stats",
        "tags": [
          "PlayerStats"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "PlayerStats ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerStats"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player stats updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Player stats not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Partially update player stats",
        "tags": [
          "PlayerStats"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "PlayerStats ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appearances": {
                    "type": "number"
                  },
                  "minutesPlayed": {
                    "type": "number"
                  },
                  "goals": {
                    "type": "object",
                    "properties": {
                      "total": {
                        "type": "number"
                      },
                      "penalties": {
                        "type": "number"
                      },
                      "freeKicks": {
                        "type": "number"
                      }
                    }
                  },
                  "assists": {
                    "type": "number"
                  },
                  "tackles": {
                    "type": "number"
                  },
                  "interceptions": {
                    "type": "number"
                  },
                  "clearances": {
                    "type": "number"
                  },
                  "cleanSheets": {
                    "type": "number"
                  },
                  "saves": {
                    "type": "number"
                  },
                  "cards": {
                    "type": "object",
                    "properties": {
                      "yellow": {
                        "type": "number"
                      },
                      "red": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player stats updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/PlayerStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Player stats not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete player stats",
        "tags": [
          "PlayerStats"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "PlayerStats ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Player stats deleted"
          },
          "404": {
            "description": "Player stats not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlayerName": {
        "type": "object",
        "required": [
          "first",
          "last"
        ],
        "properties": {
          "first": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Player's first name"
          },
          "last": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Player's last name"
          },
          "displayName": {
            "type": "string",
            "maxLength": 100,
            "description": "Player's display name (optional)"
          }
        }
      },
      "MarketValue": {
        "type": "object",
        "required": [
          "value",
          "currency"
        ],
        "properties": {
          "value": {
            "type": "number",
            "minimum": 0,
            "description": "Player's market value"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "GBP",
              "USD"
            ],
            "description": "Currency of the market value"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when market value was set"
          }
        }
      },
      "Player": {
        "type": "object",
        "required": [
          "name",
          "position",
          "nationality",
          "dateOfBirth",
          "marketValue"
        ],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/PlayerName"
          },
          "position": {
            "type": "string",
            "enum": [
              "GK",
              "DEF",
              "MID",
              "FWD"
            ],
            "description": "Player's position"
          },
          "nationality": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Player's nationality"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "Player's birth date (must be between 15-45 years old)"
          },
          "height": {
            "type": "number",
            "minimum": 150,
            "maximum": 220,
            "description": "Player's height in cm"
          },
          "weight": {
            "type": "number",
            "minimum": 50,
            "maximum": 120,
            "description": "Player's weight in kg"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INJURED",
              "SUSPENDED",
              "ON_LOAN",
              "INACTIVE"
            ],
            "default": "ACTIVE",
            "description": "Player's current status"
          },
          "jerseyNumber": {
            "type": "number",
            "minimum": 1,
            "maximum": 99,
            "description": "Player's jersey number (must be unique)"
          },
          "marketValue": {
            "$ref": "#/components/schemas/MarketValue"
          },
          "currentContract": {
            "type": "string",
            "format": "objectId",
            "description": "Reference to current Contract model"
          },
          "stats": {
            "type": "string",
            "format": "objectId",
            "description": "Reference to PlayerStats model"
          }
        }
      },
...
      "PlayerStats": {
        "type": "object",
        "required": [
          "appearances",
          "minutesPlayed"
        ],
        "properties": {
          "appearances": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of appearances made"
          },
          "minutesPlayed": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Total minutes played"
          },
          "goals": {
            "$ref": "#/components/schemas/Goals"
          },
          "assists": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of assists provided"
          },
          "tackles": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of successful tackles"
          },
          "interceptions": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of interceptions made"
          },
          "clearances": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of clearances made"
          },
          "cleanSheets": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of clean sheets kept"
          },
          "saves": {
            "type": "number",
            "minimum": 0,
            "default": 0,
            "description": "Number of saves made"
          },
          "cards": {
            "$ref": "#/components/schemas/Cards"
          }
        }
      },
...
    },
    "responses": {
...
      "PlayerResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Player"
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "string"
              },
              "collection": {
                "type": "string"
              }
            }
          }
        }
      },
      "PlayerStatsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlayerStats"
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "string"
              },
              "collection": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}